try_llm/requirements.txt
openai
python-dotenv
ollama
streamlit
langchain 
gradio 
duckduckgo-search